+++ tasty-plutus/CHANGELOG.md
@@ -4,6 +4,12 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
+## 4.0 -- 2021-11-11
+
+- Plutus upgraded
+  - `plutus` pinned to `3f089ccf0ca746b399c99afe51e063b0640af547`
+  - `plutus-apps` pinned to `404af7ac3e27ebcb218c05f79d9a70ca966407c9`
+
 ## 3.4 -- 2021-11-03
 
 * Add `ScriptInputPosition` option to control where in `txInfoInputs` the script